+++ man-roxygen/args-dots.R
@@ -1,4 +1,5 @@
 #' @param quantiles For dot plots, an optional integer passed to
 #'   [ggdist::stat_dots()] specifying the number of quantiles to use for a
-#'   quantile dot plot. If `quantiles` is `NA` (the default) then all data
-#'   points are plotted.
+#'   quantile dot plot. If `quantiles` is `NA` then all data points are plotted.
+#'   The default is `quantiles=100` so that each dot represent one percent of
+#'   posterior mass.