+++ NEWS.md
@@ -1,5 +1,8 @@
 # bayesplot (development version)
 
+* New functions `mcmc_dots` and `mcmc_dots_by_chain` for dot plots of MCMC draws by @behramulukir (#402)
+* Default to `quantiles=100` for all dot plots by @behramulukir (#402)
+
 # bayesplot 1.15.0
 
 * Add `shape` argument to `mcmc_scatter` by @behramulukir (#375)